+++ utils/cautils/offline.go
@@ -190,8 +190,9 @@ func (c *OfflineCA) Sign(req *api.SignRequest) (*api.SignResponse, error) {
 		return nil, err
 	}
 	signOpts := provisioner.SignOptions{
-		NotBefore: req.NotBefore,
-		NotAfter:  req.NotAfter,
+		NotBefore:    req.NotBefore,
+		NotAfter:     req.NotAfter,
+		TemplateData: req.TemplateData,
 	}
 	certChain, err := c.authority.Sign(req.CsrPEM.CertificateRequest, signOpts, opts...)
 	if err != nil {
@@ -531,6 +532,8 @@ func (c *OfflineCA) GenerateToken(ctx *cli.Context, tokType int, subject string,
 		return generateX5CToken(ctx, p, tokType, tokAttrs)
 	case *provisioner.SSHPOP: // Generate an SSHPOP token using ssh cert + key.
 		return generateSSHPOPToken(ctx, p, tokType, tokAttrs)
+	case *provisioner.Nebula: // Generate an JWK with an nebula header and signature.
+		return generateNebulaToken(ctx, p, tokType, tokAttrs)
 	case *provisioner.K8sSA: // Get the Kubernetes service account token.
 		return generateK8sSAToken(ctx, p)
 	case *provisioner.GCP: // Do the identity request to get the token.